refactor(eip8130): extract engine-neutral base-common-eip8130 crate
The EIP-8130 intrinsic-gas schedule/computation and the 2D nonce-manager
slot derivation were pure functions of the transaction and ERC-7201
layout, yet lived in revm-coupled crates (base-execution-eip8130,
base-common-precompiles). base-common-evm2, which must stay revm-free,
re-implemented both as "revm-free ports" kept in sync only by parity
tests — a second full copy of a ~600-line gas model that could silently
drift.

Extract them into a new no_std base-common-eip8130 crate (deps:
alloy-primitives + base-common-consensus) as the single source of truth:

- Move Eip8130GasSchedule + IntrinsicGas/IntrinsicGasInput into the new
  crate; base-execution-eip8130 re-exports them, so its executor and
  base-common-evm are unchanged.
- Add NonceManagerSlots (address, ERC-7201 base slots, mapping-slot
  derivation) with a self-contained test recomputing the slots from the
  namespace, replacing the cross-crate parity tests.
- base-common-precompiles' NonceManagerStorage and base-common-evm2's
  NonceManager now consume the shared slots; the precompile's
  handler-chain tests pin the macro storage layout to them.
- Delete base-common-evm2's duplicated gas port and slot math (and the
  now-redundant parity tests); it depends on the shared crate directly,
  keeping only its evm2-specific read/write access.

Relocate the two guard tests that need the revm path (revm gas-constant
drift tripwire; apply-vs-intrinsic policy-decode coupling) into a
base-execution-eip8130 integration test.
